% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphExperiment-methods.R
\name{GraphExperiment-methods}
\alias{GraphExperiment-methods}
\alias{graphs}
\alias{graphs<-}
\alias{graph}
\alias{graph<-}
\alias{graphNames}
\alias{graphNames<-}
\alias{graphs,GraphExperiment-method}
\alias{graph,GraphExperiment,missing-method}
\alias{graph,GraphExperiment,ANY-method}
\alias{graphNames,GraphExperiment-method}
\alias{graphs<-,GraphExperiment-method}
\alias{graph<-,GraphExperiment-method}
\alias{graphNames<-,GraphExperiment,character-method}
\title{Methods for \code{GraphExperiment} objects}
\usage{
\S4method{graphs}{GraphExperiment}(x)

\S4method{graph}{GraphExperiment,missing}(x, i)

\S4method{graph}{GraphExperiment,ANY}(x, i)

\S4method{graphNames}{GraphExperiment}(x)

\S4method{graphs}{GraphExperiment}(x) <- value

\S4method{graph}{GraphExperiment}(x, i) <- value

\S4method{graphNames}{GraphExperiment,character}(x) <- value
}
\arguments{
\item{x}{A \code{GraphExperiment} object.}

\item{i}{List element (numeric for index, character for name) of the element
to access or replace.}

\item{value}{Replacement value for replacement methods}
}
\value{
Return values depend on the method. See details and examples.
}
\description{
The \code{GraphExperiment} class provides users with methods to get and
set graphs (\code{igraph} objects) representing how features of
\code{\linkS4class{SingleCellExperiment}} objects relate to
each other.
}
\section{graphs and graph methods}{

\describe{
\item{\code{graphs(x)}: }{
Getter for a \code{SimpleList} of \code{igraph} objects.}
\item{\code{graphs(x) <- value}: }{
Setter for a SimpleList or list (coerced to SimpleList) of \code{igraph}
objects.}
\item{\code{graph(x, i)}: }{
Getter for an \code{igraph} object containing graph \code{i} from the list
stored in \code{graphs}.}
\item{\code{graph(x, i) <- value}: }{
Setter for an \code{igraph} object to be stored in element \code{i} of the
list \code{graphs}}
}
}

\section{graphNames methods}{

\describe{
\item{\code{graphNames(x)}: }{
Getter to extract names of graphs in the \code{graphs} slot.}
\item{\code{graphNames(x) <-  value}: }{
Setter to assign new names to the graphs stored in the \code{graphs}
slot.}
}
}

\examples{
# Create a GraphExperiment object
gene_ids <- paste0("gene", seq_len(200))
cell_ids <- paste0("cell", seq_len(100))
mat <- matrix(rpois(20000, 5), ncol = 100, dimnames = list(gene_ids, cell_ids))
g <- graph_from_adjacency_matrix(cor(t(mat)), weighted = TRUE)

ge <- GraphExperiment(assays = list(counts = mat), graphs = list(cor = g))
ge

# Extract graph names
graphNames(ge)
ge

# Extract graphs
graphs(ge)
graph(ge, "cor")

# Add a new graph
graph(ge, "newcor") <- g
ge

# Add a list of graphs
graphs(ge) <- list(cor = g, newcor = g)
ge

# Replace graph names
graphNames(ge) <- c("network1", "network2")
ge

}
